ml/requirements.txt
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
joblib==1.3.1
# 開発・品質管理用ツール
flake8==6.0.0
black==23.7.0
isort==5.12.0